+++ CHANGELOG.md
@@ -7,6 +7,7 @@ NOTE: isort follows the [semver](https://semver.org/) versioning standard.
   - Implemented ability to treat all or select comments as code (issue #1357)
   - Implemented ability to use different configs for different file extensions (issue #1162)
   - Added experimental support for sorting literals (issue #1358)
+  - Added experimental support for sorting and deduping groupings of assignments.
   - Improved handling of deprecated single line variables for usage with Visual Studio Code (issue #1363)
   - Improved handling of mixed newline forms within same source file.
   - Improved error handling for known sections.